requirements.txt
pyTelegramBotApi
flask
gunicorn
aiogram
aiohttp
lxml